# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "WebCheckly 网站检测服务 API 文档",
        "title": "WebCheckly API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@webcheckly.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/scan": {
            "get": {
                "description": "使用Server-Sent Events进行实时流式扫描。此接口已废弃，建议使用任务管理API（POST /api/scans）。\n\nSSE事件类型：\n- start: 扫描开始\n- result: 单个链接检测结果\n- progress: 进度更新\n- website-info: 网站信息\n- domain-info: 域名信息\n- ssl-info: SSL证书信息\n- tech-stack: 技术栈信息\n- performance-info: 性能指标\n- seo-info: SEO合规性\n- security-info: 安全风险\n- accessibility-info: 可访问性\n- ai-analysis: AI分析报告\n- done: 扫描完成\n- error: 错误信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "扫描"
                ],
                "summary": "SSE扫描接口（降级方案）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "目标URL",
                        "name": "url",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "扫描选项",
                        "name": "options",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "zh",
                            "en"
                        ],
                        "type": "string",
                        "default": "zh",
                        "description": "语言",
                        "name": "lang",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "performance",
                            "security",
                            "seo",
                            "balanced"
                        ],
                        "type": "string",
                        "default": "balanced",
                        "description": "AI分析模式",
                        "name": "ai_mode",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SSE事件流（text/event-stream格式）",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "请求参数错误（URL格式错误、私有IP等）",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "429": {
                        "description": "请求过于频繁（触发限流）",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "服务繁忙（并发任务数已达上限）",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/scans": {
            "post": {
                "description": "创建一个新的网站扫描任务，支持多种检测选项。任务创建后会异步执行，可通过任务ID查询状态和结果。\n\n支持的扫描选项：\n- website-info: 网站基础信息（标题、描述、关键词等）\n- domain-info: 域名DNS信息（IP、MX、NS、TXT记录等）\n- ssl-info: SSL证书信息（有效期、签名算法等）\n- tech-stack: 技术栈识别（框架、CMS、CDN等）\n- link-health: 链接健康检查（检测页面内所有链接的可用性）\n- performance: 性能检测（Lighthouse性能指标）\n- seo: SEO合规性检测（Lighthouse SEO指标）\n- security: 安全风险检测（Lighthouse安全指标）\n- accessibility: 可访问性检测（Lighthouse A11y指标）\n- ai-analysis: AI智能分析报告（需要配置DEEPSEEK_API_KEY）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务管理"
                ],
                "summary": "创建扫描任务",
                "parameters": [
                    {
                        "description": "任务创建请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "任务创建成功，返回任务ID和初始状态",
                        "schema": {
                            "$ref": "#/definitions/models.CreateTaskResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误（URL格式错误、私有IP等）",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/scans/{id}": {
            "get": {
                "description": "查询指定任务的执行状态和进度信息，包括各模块的执行状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务管理"
                ],
                "summary": "获取任务状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "任务状态",
                        "schema": {
                            "$ref": "#/definitions/models.TaskStatusResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "任务不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateTaskRequest": {
            "description": "创建扫描任务的请求参数",
            "type": "object",
            "properties": {
                "ai_mode": {
                    "description": "AI分析模式",
                    "type": "string",
                    "default": "balanced",
                    "enum": [
                        "performance",
                        "security",
                        "seo",
                        "balanced"
                    ],
                    "example": "balanced"
                },
                "language": {
                    "description": "语言 (zh/en)",
                    "type": "string",
                    "default": "zh",
                    "enum": [
                        "zh",
                        "en"
                    ],
                    "example": "zh"
                },
                "options": {
                    "description": "扫描选项",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "website-info",
                        "domain-info"
                    ]
                },
                "url": {
                    "description": "目标URL",
                    "type": "string",
                    "example": "https://example.com"
                }
            }
        },
        "models.CreateTaskResponse": {
            "description": "创建任务成功后的响应",
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "status": {
                    "enum": [
                        "pending",
                        "running",
                        "completed",
                        "failed"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.TaskStatus"
                        }
                    ],
                    "example": "pending"
                }
            }
        },
...
        "models.ModuleStatus": {
            "description": "单个检测模块的执行状态",
            "type": "object",
            "properties": {
                "completed_at": {
                    "description": "完成时间",
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "error": {
                    "description": "错误信息（如果有）",
                    "type": "string",
                    "example": ""
                },
                "name": {
                    "description": "模块名称",
                    "type": "string",
                    "example": "website-info"
                },
                "progress": {
                    "description": "模块进度",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.TaskProgress"
                        }
                    ]
                },
                "started_at": {
                    "description": "开始时间",
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "status": {
                    "description": "模块状态",
                    "enum": [
                        "pending",
                        "running",
                        "completed",
                        "failed"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.TaskStatus"
                        }
                    ],
                    "example": "running"
                }
            }
        },
...
        "models.TaskProgress": {
            "description": "任务执行进度信息",
            "type": "object",
            "properties": {
                "current": {
                    "description": "当前进度",
                    "type": "integer",
                    "example": 5
                },
                "total": {
                    "description": "总进度",
                    "type": "integer",
                    "example": 10
                }
            }
        },
...
        "models.TaskStatus": {
            "type": "string",
            "enum": [
                "pending",
                "running",
                "completed",
                "failed"
            ],
            "x-enum-comments": {
                "TaskStatusCompleted": "已完成",
                "TaskStatusFailed": "执行失败",
                "TaskStatusPending": "等待执行",
                "TaskStatusRunning": "执行中"
            },
            "x-enum-descriptions": [
                "等待执行",
                "执行中",
                "已完成",
                "执行失败"
            ],
            "x-enum-varnames": [
                "TaskStatusPending",
                "TaskStatusRunning",
                "TaskStatusCompleted",
                "TaskStatusFailed"
            ]
        },
        "models.TaskStatusResponse": {
            "description": "任务状态查询响应",
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "error": {
                    "type": "string",
                    "example": ""
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "modules": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/models.ModuleStatus"
                    }
                },
                "progress": {
                    "$ref": "#/definitions/models.TaskProgress"
                },
                "status": {
                    "enum": [
                        "pending",
                        "running",
                        "completed",
                        "failed"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.TaskStatus"
                        }
                    ],
                    "example": "running"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "可选：Bearer token认证（当前版本暂未启用）",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}